AYH/requirements.txt
# Django Core
Django==5.2.10

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# CORS Headers (for Flutter/Mobile apps)
django-cors-headers==4.3.1

# Additional utilities
python-decouple==3.8  # For environment variables
Pillow==10.1.0  # For image handling (if needed later)

# Production server (optional)
gunicorn==21.2.0
whitenoise==6.6.0  # Static files serving
